Updated the contract for a monthly-fee based offer
This is an update of Sam's version of the Contract Killer but adjusted it to fit the business model of redditor Citrous_Oyster.

This means it's for an offer of a static website that focuses on performance and search engine rankings. It includes ongoing hosting and maintenance, and all changes and updates the client wants will be made by the developer, while the client will pay a monthly fee with no money down at the start.